+++ src/main/java/coupon/common/infra/datasource/DataSourceHelper.java
@@ -1,18 +1,19 @@
-package coupon.infra.datasource;
+package coupon.common.infra.datasource;
 
 import java.util.function.Supplier;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Component
 public class DataSourceHelper {
 
-    @Transactional
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     public void executeInWriter(Runnable runnable) {
         runnable.run();
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.REQUIRES_NEW)
     public <T> T executeInWriter(Supplier<T> supplier) {
         return supplier.get();
     }
