+++ src/main/java/coupon/coupon/domain/CouponCategory.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.coupon.domain;
 
 public enum CouponCategory {
