+++ src/main/java/coupon/coupon/domain/CouponDiscountAmount.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.coupon.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
