+++ src/main/java/coupon/coupon/domain/CouponIssueDate.java
@@ -1,4 +1,4 @@
-package coupon.domain;
+package coupon.coupon.domain;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
