+++ src/main/java/coupon/membercoupon/domain/MemberCoupon.java
@@ -1,6 +1,7 @@
-package coupon.domain;
+package coupon.membercoupon.domain;
 
 import java.time.LocalDateTime;
+import coupon.coupon.domain.Coupon;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
