+++ src/test/java/coupon/coupon/application/CouponServiceTest.java
@@ -1,11 +1,12 @@
-package coupon.application;
+package coupon.coupon.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
-import coupon.domain.Coupon;
-import coupon.domain.CouponCategory;
-import coupon.infra.datasource.DataSourceHelper;
+import coupon.coupon.application.CouponService;
+import coupon.coupon.domain.Coupon;
+import coupon.coupon.domain.CouponCategory;
+import coupon.common.infra.datasource.DataSourceHelper;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
