+++ src/test/java/coupon/coupon/domain/CouponDiscountAmountTest.java
@@ -1,8 +1,10 @@
-package coupon.domain;
+package coupon.coupon.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import coupon.coupon.domain.CouponDiscountAmount;
+import coupon.coupon.domain.CouponMinOrderAmount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
