+++ src/test/java/coupon/coupon/domain/CouponNameTest.java
@@ -1,8 +1,9 @@
-package coupon.domain;
+package coupon.coupon.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import coupon.coupon.domain.CouponName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
