+++ src/test/java/coupon/coupon/domain/CouponTest.java
@@ -1,8 +1,10 @@
-package coupon.domain;
+package coupon.coupon.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.time.LocalDate;
+import coupon.coupon.domain.Coupon;
+import coupon.coupon.domain.CouponCategory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
